Code formatting, minor improvements

+++ src/bsdfs/microfacet.hpp
@@ -16,7 +16,7 @@ namespace lightwave::microfacet {
  * @c RoughDielectric ).
  * @see https://www.graphics.cornell.edu/~bjw/microfacetbsdf.pdf (Figure 6)
  */
-inline float detReflection(const Vector &normal, const Vector &wo) {
+inline float detReflection(const Vector& normal, const Vector& wo) {
     return 1 / abs(4 * normal.dot(wo));
 }
 
@@ -27,8 +27,7 @@ inline float detReflection(const Vector &normal, const Vector &wo) {
  * @c RoughDielectric ).
  * @see https://www.graphics.cornell.edu/~bjw/microfacetbsdf.pdf (Figure 7)
  */
-inline float detRefraction(const Vector &normal, const Vector &wi,
-                           const Vector &wo, float eta) {
+inline float detRefraction(const Vector& normal, const Vector& wi, const Vector& wo, float eta) {
     return sqr(eta) *
            abs(normal.dot(wi) / sqr(normal.dot(wi) * eta + normal.dot(wo)));
 }
@@ -39,7 +38,7 @@ inline float detRefraction(const Vector &normal, const Vector &wi,
  * @param wh The sampled normal (also known as half vector)
  * @param w The direction to evaluate for (either @c wo or @c wi )
  */
-inline float smithG1(float alpha, const Vector &wh, const Vector &w) {
+inline float smithG1(float alpha, const Vector& wh, const Vector& w) {
     /// Ensure correct orientation by projecting both @c w and @c wh into the
     /// upper hemisphere and checking that the angle they form is less than 90°
     if (w.dot(wh) * Frame::cosTheta(w) * Frame::cosTheta(wh) <= 0)
@@ -61,11 +60,11 @@ inline float smithG1(float alpha, const Vector &wh, const Vector &w) {
  * @see "Microfacet Models for Refraction through Rough Surfaces" [Walter et al.
  * 2007]
  */
-inline float evaluateGGX(float alpha, const Vector &wh) {
+inline float evaluateGGX(float alpha, const Vector& wh) {
     float nDotH = Frame::cosTheta(wh);
-    float a     = Frame::cosPhiSinTheta(wh) / alpha;
-    float b     = Frame::sinPhiSinTheta(wh) / alpha;
-    float c     = sqr(a) + sqr(b) + sqr(nDotH);
+    float a = Frame::cosPhiSinTheta(wh) / alpha;
+    float b = Frame::sinPhiSinTheta(wh) / alpha;
+    float c = sqr(a) + sqr(b) + sqr(nDotH);
     return 1 / (Pi * sqr(alpha * c));
 }
 
@@ -80,31 +79,28 @@ inline float evaluateGGX(float alpha, const Vector &wh) {
  * @see For details on how and why this works, check out Eric Heitz' great JCGT
  * paper "Sampling the GGX Distribution of Visible Normals".
  */
-inline Vector sampleGGXVNDF(float alpha, const Vector &wo, const Point2 &rnd) {
+inline Vector sampleGGXVNDF(float alpha, const Vector& wo, const Point2& rnd) {
     // Addition: flip sign of incident vector for transmission
     float sgn = copysign(1, Frame::cosTheta(wo));
     // Section 3.2: transforming the view direction to the hemisphere
     // configuration
-    Vector Vh =
-        sgn * Vector(alpha * wo.x(), alpha * wo.y(), wo.z()).normalized();
+    Vector Vh = sgn * Vector(alpha * wo.x(), alpha * wo.y(), wo.z()).normalized();
     // Section 4.1: orthonormal basis (with special case if cross product is
     // zero)
     float lensq = Vh.x() * Vh.x() + Vh.y() * Vh.y();
-    Vector T1 =
-        lensq > 0 ? Vector(-Vh.y(), Vh.x(), 0) / sqrt(lensq) : Vector(1, 0, 0);
+    Vector T1 = lensq > 0 ? Vector(-Vh.y(), Vh.x(), 0) / sqrt(lensq) : Vector(1, 0, 0);
     Vector T2 = Vh.cross(T1);
     // Section 4.2: parameterization of the projected area
-    float r   = sqrt(rnd.x());
+    float r = sqrt(rnd.x());
     float phi = 2 * Pi * rnd.y();
-    float t1  = r * cos(phi);
-    float t2  = r * sin(phi);
-    float s   = 0.5f * (1 + Vh.z());
-    t2        = (1 - s) * sqrt(1 - sqr(t1)) + s * t2;
+    float t1 = r * cosf(phi);
+    float t2 = r * sinf(phi);
+    float s = 0.5f * (1 + Vh.z());
+    t2 = (1 - s) * sqrt(1 - sqr(t1)) + s * t2;
     // Section 4.3: reprojection onto hemisphere
     Vector Nh = t1 * T1 + t2 * T2 + safe_sqrt(1 - sqr(t1) - sqr(t2)) * Vh;
     // Section 3.4: transforming the normal back to the ellipsoid configuration
-    Vector Ne =
-        Vector(alpha * Nh.x(), alpha * Nh.y(), max(0, Nh.z())).normalized();
+    Vector Ne = Vector(alpha * Nh.x(), alpha * Nh.y(), max(0, Nh.z())).normalized();
     return sgn * Ne;
 }
 
@@ -115,12 +111,10 @@ inline Vector sampleGGXVNDF(float alpha, const Vector &wo, const Point2 &rnd) {
  * @param wo The outgoing direction (determines which normals are visible)
  * @returns The probability density of @c sampleGGXVNDF sampling @c wh .
  */
-inline float pdfGGXVNDF(float alpha, const Vector &wh, const Vector &wo) {
-    // clang-format off
+inline float pdfGGXVNDF(float alpha, const Vector& wh, const Vector& wo) {
     return microfacet::evaluateGGX(alpha, wh) *
            microfacet::smithG1(alpha, wh, wo) *
            abs(wh.dot(wo)) / Frame::absCosTheta(wo);
-    // clang-format on
 }
 
 //
@@ -139,8 +133,7 @@ inline float pdfGGXVNDF(float alpha, const Vector &wh, const Vector &wo) {
  * @param wh The sampled normal (also known as half vector)
  * @param w The direction to evaluate for (either @c wo or @c wi )
  */
-inline float anisotropicSmithG1(float ax, float ay, const Vector &wh,
-                                const Vector &w) {
+inline float anisotropicSmithG1(float ax, float ay, const Vector& wh, const Vector& w) {
     /// Ensure correct orientation by projecting both @c w and @c wh into the
     /// upper hemisphere and checking that the angle they form is less than 90°
     if (w.dot(wh) * Frame::cosTheta(w) * Frame::cosTheta(wh) <= 0)
@@ -165,11 +158,11 @@ inline float anisotropicSmithG1(float ax, float ay, const Vector &wh,
  * @see "Microfacet Models for Refraction through Rough Surfaces" [Walter et al.
  * 2007]
  */
-inline float evaluateAnisotropicGGX(float ax, float ay, const Vector &wh) {
+inline float evaluateAnisotropicGGX(float ax, float ay, const Vector& wh) {
     float nDotH = Frame::cosTheta(wh);
-    float a     = Frame::cosPhiSinTheta(wh) / ax;
-    float b     = Frame::sinPhiSinTheta(wh) / ay;
-    float c     = sqr(a) + sqr(b) + sqr(nDotH);
+    float a = Frame::cosPhiSinTheta(wh) / ax;
+    float b = Frame::sinPhiSinTheta(wh) / ay;
+    float c = sqr(a) + sqr(b) + sqr(nDotH);
     return 1 / (Pi * ax * ay * sqr(c));
 }
 
@@ -185,8 +178,7 @@ inline float evaluateAnisotropicGGX(float ax, float ay, const Vector &wh) {
  * @see For details on how and why this works, check out Eric Heitz' great JCGT
  * paper "Sampling the GGX Distribution of Visible Normals".
  */
-inline Vector sampleAnisotropicGGXVNDF(float ax, float ay, const Vector &wo,
-                                       const Point2 &rnd) {
+inline Vector sampleAnisotropicGGXVNDF(float ax, float ay, const Vector& wo, const Point2& rnd) {
     // Addition: flip sign of incident vector for transmission
     float sgn = copysign(1, Frame::cosTheta(wo));
     // Section 3.2: transforming the view direction to the hemisphere
@@ -195,16 +187,15 @@ inline Vector sampleAnisotropicGGXVNDF(float ax, float ay, const Vector &wo,
     // Section 4.1: orthonormal basis (with special case if cross product is
     // zero)
     float lensq = Vh.x() * Vh.x() + Vh.y() * Vh.y();
-    Vector T1 =
-        lensq > 0 ? Vector(-Vh.y(), Vh.x(), 0) / sqrt(lensq) : Vector(1, 0, 0);
+    Vector T1 = lensq > 0 ? Vector(-Vh.y(), Vh.x(), 0) / sqrt(lensq) : Vector(1, 0, 0);
     Vector T2 = Vh.cross(T1);
     // Section 4.2: parameterization of the projected area
-    float r   = sqrt(rnd.x());
+    float r = sqrt(rnd.x());
     float phi = 2 * Pi * rnd.y();
-    float t1  = r * cos(phi);
-    float t2  = r * sin(phi);
-    float s   = 0.5f * (1 + Vh.z());
-    t2        = (1 - s) * sqrt(1 - sqr(t1)) + s * t2;
+    float t1 = r * cosf(phi);
+    float t2 = r * sinf(phi);
+    float s = 0.5f * (1 + Vh.z());
+    t2 = (1 - s) * sqrt(1 - sqr(t1)) + s * t2;
     // Section 4.3: reprojection onto hemisphere
     Vector Nh = t1 * T1 + t2 * T2 + safe_sqrt(1 - sqr(t1) - sqr(t2)) * Vh;
     // Section 3.4: transforming the normal back to the ellipsoid configuration
@@ -221,8 +212,7 @@ inline Vector sampleAnisotropicGGXVNDF(float ax, float ay, const Vector &wo,
  * @param wo The outgoing direction (determines which normals are visible)
  * @returns The probability density of @c sampleGGXVNDF sampling @c wh .
  */
-inline float pdfAnisotropicGGXVNDF(float ax, float ay, const Vector &normal,
-                                   const Vector &wo) {
+inline float pdfAnisotropicGGXVNDF(float ax, float ay, const Vector& normal, const Vector& wo) {
     return microfacet::evaluateAnisotropicGGX(ax, ay, normal) *
            microfacet::anisotropicSmithG1(ax, ay, normal, wo) *
            abs(normal.dot(wo)) / Frame::absCosTheta(wo);
@@ -247,11 +237,11 @@ inline float pdfAnisotropicGGXVNDF(float ax, float ay, const Vector &normal,
  * @see "Diffuse Reflection of Light from a Matt Surface" [Berry 1923]
  * @see "Physically Based Shading at Disney" [Burley 2012]
  */
-inline float evaluateGTR1(float alpha, const Vector &wh) {
+inline float evaluateGTR1(float alpha, const Vector& wh) {
     float nDotH = Frame::cosTheta(wh);
-    float a2    = sqr(alpha);
-    float t     = 1 + (a2 - 1) * sqr(nDotH);
-    return (a2 - 1) / (Pi * log(a2) * t);
+    float a2 = sqr(alpha);
+    float t = 1 + (a2 - 1) * sqr(nDotH);
+    return (a2 - 1) / (Pi * logf(a2) * t);
 }
 
 /**
@@ -262,16 +252,16 @@ inline float evaluateGTR1(float alpha, const Vector &wh) {
  * @note The PDF of @c wh is given by:
  *   @code cosTheta(wh) * D(wh) @endcode
  */
-inline Vector sampleGTR1(float alpha, const Point2 &rnd) {
+inline Vector sampleGTR1(float alpha, const Point2& rnd) {
     float a2 = sqr(alpha);
 
-    float cosTheta = safe_sqrt((1 - pow(a2, 1 - rnd.x())) / (1 - a2));
+    float cosTheta = safe_sqrt((1 - powf(a2, 1 - rnd.x())) / (1 - a2));
     float sinTheta = safe_sqrt(1 - (cosTheta * cosTheta));
-    float phi      = 2 * Pi * rnd.y();
-    float sinPhi   = sin(phi);
-    float cosPhi   = cos(phi);
+    float phi = 2 * Pi * rnd.y();
+    float sinPhi = sinf(phi);
+    float cosPhi = cosf(phi);
 
-    return { sinTheta * cosPhi, sinTheta * sinPhi, cosTheta };
+    return {sinTheta * cosPhi, sinTheta * sinPhi, cosTheta};
 }
 
 } // namespace lightwave::microfacet
